Add docs and reorg vllm_fp8_utils.py
